requirements.txt
#tensorflow==2.12.0
#langchain-community==0.0.3
#tf-keras==2.12.0
langchain==0.0.184
python-dotenv==1.0.0
streamlit==1.18.1
faiss-cpu==1.7.4
huggingface-hub==0.14.1 
InstructorEmbedding==1.0.1
sentence-transformers==2.2.2
torch==2.0.1
altair==4.2.2
